remove reducer folder
