+++ src/stores/randomQuoteStore.ts
@@ -1,13 +1,60 @@
 import { useReducer } from 'react';
 
-import { RANDOM_QUOTE_ERROR, RANDOM_QUOTE_SUCCESS } from 'src/actions/types';
+import {
+  RANDOM_QUOTE_ERROR,
+  RANDOM_QUOTE_INIT,
+  RANDOM_QUOTE_LOADING,
+  RANDOM_QUOTE_SUCCESS,
+  RandomQuoteActionType,
+} from 'src/actions/types';
 import restApiRandomQuote from 'src/http/restApiRandomQuote';
-import { randomQuoteReducer, RandomQuoteState } from 'src/reducers/randomQuoteReducer';
+import Quote from 'src/models/Quote';
+
+export interface RandomQuoteState {
+  loading: boolean;
+  data: Quote | null;
+  error: Error | null | unknown;
+}
 
 export interface RandomQuoteStore {
   state: RandomQuoteState;
   fetchCharacterRandomQuote: (characterName: string) => Promise<void>;
 }
+
+export const randomQuoteReducer = (state: RandomQuoteState, action: RandomQuoteActionType): RandomQuoteState => {
+  switch (action.type) {
+    case RANDOM_QUOTE_INIT:
+      return {
+        loading: false,
+        data: null,
+        error: null,
+      };
+    case RANDOM_QUOTE_LOADING:
+      return {
+        loading: true,
+        data: null,
+        error: null,
+      };
+    case RANDOM_QUOTE_SUCCESS:
+      return {
+        loading: false,
+        data: action.data,
+        error: null,
+      };
+    case RANDOM_QUOTE_ERROR:
+      return {
+        loading: false,
+        data: null,
+        error: action.error,
+      };
+    default:
+      return {
+        loading: false,
+        data: null,
+        error: new Error('Unhandled action type'),
+      };
+  }
+};
 
 const useRandomQuoteStore = (): RandomQuoteStore => {
   const [state, dispatch] = useReducer(randomQuoteReducer, {
